Make get_request timeouts recoverable
With short select() delays (e.g. in event loops) and slow connections,
the client may not have more data ready before select timeouts.

This change lets get_request return a different kind of false (defined
instead of undef) in case of a select timeout, so that the caller can
choose to call it again. To make this work, the same buffer
is re-used with _need_more every time.

Hacks that inherit from HTTP::Daemon and override _need_more, will
break, as will existing code that explicitly tests the result of
get_request for definedness rather than its boolean value, and then
assumes that it will be an HTTP::Request object. Both cases are
unlikely.

The unit test is based on chunked.t with a lot removed.