create TestServer test lib for running daemon process
Create a TestServer test library that serves as both a subclassable
server and a lib for running that server. Having one class for both the
internals of the server and the external interface to start it is a bit
ugly, but makes extending it and using it a bit easier. And it's only a
test library so it seems a reasonable compromise.

This will run the server as a sub process rather than using forking, so
it is more portable.

Convert the tests to use this library rather than each having their own
routines for running for forking the daemon.